+++ src/react/components/Footer/index.js
@@ -1,28 +1,30 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {Footer,Text,Wrapper} from './StoryBook'
+import {Text} from '../StoryBook'
+import {FooterWrapper, Footer} from './components'
 
 
-
-const FooterComponent=({background,text,isDesktop})=>(
-        <Wrapper >
+const FooterComponent=({background,text})=>(
+        <FooterWrapper >
             <Footer bg={background}>
                 <Text textAlign={"center"} >
                     {text}
                 </Text>
             </Footer>
-        </Wrapper>
+        </FooterWrapper>
 )
 
 FooterComponent.propTypes={
     text:PropTypes.string,
     background:PropTypes.oneOfType(
-        [PropTypes.string,PropTypes.object]
+        [
+            PropTypes.string,
+            PropTypes.object
+        ]
         )
 }
 FooterComponent.defaultProps={
     text:"Footer Text",
-    background:null,
-    isDesktop:false
+    background:null
 }
 export default FooterComponent
